+++ src/Enum/SystemOwnerNoticeEnum.php
@@ -7,4 +7,5 @@ enum SystemOwnerNoticeEnum: string
     case SERVERFLYTNING = 'serverflytning';
     case SIKKERHEDSPATCH = 'sikkerhedspatch';
     case CYBERSIKKERSHEDSOPDATERING = 'cybersikkershedsopdatering';
+    case RELEASE = 'release';
 }